user interface provided
We also provided the codes of user interface to convert multiple csv files into one weka .arff file or see5 .names and .da files. 
The merging function will connect all files via attributes names (column names) and drop duplicated rows and columns with same attribute name and data. If there's a conflict of attribute names with different data, the column will be kept with a suffix added to the attribute name as the rename.